requirements.txt
absl-py==1.3.0
aiohttp==3.8.3
aiosignal==1.3.1
astunparse==1.6.3
async-timeout==4.0.2
attrs==22.1.0
boto3==1.26.7
botocore==1.29.7
brotlipy==0.7.0
cachetools==4.2.4
certifi @ file:///C:/b/abs_ac29jvt43w/croot/certifi_1665076682579/work/certifi
cffi @ file:///C:/Windows/Temp/abs_6808y9x40v/croots/recipe/cffi_1659598653989/work
chardet==3.0.4
charset-normalizer @ file:///tmp/build/80754af9/charset-normalizer_1630003229654/work
click @ file:///C:/ci/click_1646056762388/work
colorama @ file:///C:/Windows/TEMP/abs_9439aeb1-0254-449a-96f7-33ab5eb17fc8apleb4yn/croots/recipe/colorama_1657009099097/work
contourpy==1.0.6
cryptography @ file:///C:/b/abs_36x9ifdcl4/croot/cryptography_1665612655344/work
cycler==0.11.0
datasets==2.6.1
dill==0.3.5.1
evaluate==0.3.0
filelock @ file:///opt/conda/conda-bld/filelock_1647002191454/work
fire==0.4.0
fonttools==4.38.0
frozenlist==1.3.3
fsspec==2022.10.0
future @ file:///C:/ci_310/future_1642081234309/work
google-api-core==1.33.2
google-auth==1.35.0
google-auth-oauthlib==0.4.6
google-cloud-core==1.7.3
google-cloud-translate==2.0.1
googleapis-common-protos==1.56.4
grpcio==1.50.0
grpcio-status==1.48.2
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2022.11.1
httpcore==0.9.1
httpx==0.13.3
huggingface-hub==0.10.1
hyperframe==5.2.0
idna==2.10
importlib-metadata @ file:///C:/ci/importlib-metadata_1648544469310/work
jmespath==1.0.1
joblib @ file:///C:/b/abs_e60_bwl1v6/croot/joblib_1666298845728/work
kiwisolver==1.4.4
lightning-lite==1.8.0.post1
lightning-utilities==0.3.0
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib==3.6.2
mkl-fft==1.3.1
mkl-random @ file:///C:/ci_310/mkl_random_1643050563308/work
mkl-service==2.4.0
multidict==6.0.2
multiprocess==0.70.13
nltk==3.7
numpy @ file:///C:/b/abs_53f_dbvhzc/croot/numpy_and_numpy_base_1665773185489/work
oauthlib==3.2.2
packaging @ file:///tmp/build/80754af9/packaging_1637314298585/work
pandas==1.5.1
Pillow==9.3.0
protobuf==3.20.3
pyarrow==10.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser @ file:///tmp/build/80754af9/pycparser_1636541352034/work
pyOpenSSL @ file:///opt/conda/conda-bld/pyopenssl_1643788558760/work
pyparsing @ file:///C:/Users/BUILDE~1/AppData/Local/Temp/abs_7f_7lba6rl/croots/recipe/pyparsing_1661452540662/work
PySocks @ file:///C:/ci_310/pysocks_1642089375450/work
python-dateutil==2.8.2
python-dotenv==0.21.0
pytz==2022.6
PyYAML==6.0
regex @ file:///C:/ci/regex_1658258299320/work
requests @ file:///C:/ci/requests_1657735340829/work
requests-oauthlib==1.3.1
responses==0.18.0
rfc3986==1.5.0
rsa==4.9
s3transfer==0.6.0
sacremoses @ file:///tmp/build/80754af9/sacremoses_1633107328213/work
scikit-learn==1.1.3
scipy==1.9.3
seqeval==1.2.2
six @ file:///tmp/build/80754af9/six_1644875935023/work
sniffio==1.3.0
tensorboard==2.11.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
termcolor==2.1.0
threadpoolctl==3.1.0
tokenizers @ file:///C:/ci/tokenizers_1651821358528/work
torch==1.13.0
torchaudio==0.13.0
torchmetrics==0.10.2
torchvision==0.14.0
tqdm @ file:///C:/b/abs_0axbz66qik/croots/recipe/tqdm_1664392691071/work
transformers @ file:///C:/b/abs_acffqr7197/croot/transformers_1667833417898/work
typing_extensions @ file:///C:/Windows/TEMP/abs_dd2d0moa85/croots/recipe/typing_extensions_1659638831135/work
urllib3 @ file:///C:/b/abs_a8_3vfznn_/croot/urllib3_1666298943664/work
Werkzeug==2.2.2
win-inet-pton @ file:///C:/ci_310/win_inet_pton_1642658466512/work
wincertstore==0.2
xxhash==3.1.0
yarl==1.8.1
zipp @ file:///C:/ci/zipp_1652255909324/work
